Support AIE.token in the main flow
