+++ include/aie/AIEConversionPatterns.h
@@ -1,0 +1,37 @@
+//===- AIEConversionPatterns.h ----------------------------------*- C++ -*-===//
+//
+// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+// (c) Copyright 2022 Xilinx Inc.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef MLIR_AIE_CONVERSION_PATTERNS_H
+#define MLIR_AIE_CONVERSION_PATTERNS_H
+
+#include "mlir/Transforms/DialectConversion.h"
+
+using namespace mlir;
+
+template <typename MyOp>
+struct AIEOpRemoval : public OpConversionPattern<MyOp> {
+  using OpConversionPattern<MyOp>::OpConversionPattern;
+  using OpAdaptor = typename MyOp::Adaptor;
+  ModuleOp &module;
+
+  AIEOpRemoval(MLIRContext *context, ModuleOp &m, PatternBenefit benefit = 1)
+      : OpConversionPattern<MyOp>(context, benefit), module(m) {}
+
+  LogicalResult
+  matchAndRewrite(MyOp op, OpAdaptor operands,
+                  ConversionPatternRewriter &rewriter) const override {
+    Operation *Op = op.getOperation();
+
+    rewriter.eraseOp(Op);
+    return success();
+  }
+};
+
+#endif
